backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
xgboost==2.0.1
lightgbm==4.1.0
shap==0.43.0
lime==0.2.0.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0
faker==20.1.0
